+++ dms-service/src/main/java/de/muenchen/oss/swim/dms/adapter/in/streaming/DmsEventDTO.java
@@ -1,0 +1,7 @@
+package de.muenchen.oss.swim.dms.adapter.in.streaming;
+
+public record DmsEventDTO(
+        String useCase,
+        String presignedUrl,
+        String metadataPresignedUrl) {
+}
