+++ dms-service/src/main/java/de/muenchen/oss/swim/dms/configuration/SwimDmsProperties.java
@@ -1,0 +1,20 @@
+package de.muenchen.oss.swim.dms.configuration;
+
+import de.muenchen.oss.swim.dms.domain.model.UseCase;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+import java.util.List;
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.NestedConfigurationProperty;
+import org.springframework.validation.annotation.Validated;
+
+@Data
+@ConfigurationProperties(prefix = "swim")
+@Validated
+public class SwimDmsProperties {
+    @NotNull
+    @Valid
+    @NestedConfigurationProperty
+    private List<UseCase> useCases = List.of();
+}
