+++ dms-service/src/main/java/de/muenchen/oss/swim/dms/domain/exception/DmsException.java
@@ -1,0 +1,8 @@
+package de.muenchen.oss.swim.dms.domain.exception;
+
+@SuppressWarnings("PMD.MissingSerialVersionUID")
+public class DmsException extends RuntimeException {
+    public DmsException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}
