+++ dms-service/src/main/java/de/muenchen/oss/swim/dms/domain/exception/PresignedUrlException.java
@@ -1,0 +1,12 @@
+package de.muenchen.oss.swim.dms.domain.exception;
+
+@SuppressWarnings("PMD.MissingSerialVersionUID")
+public class PresignedUrlException extends RuntimeException {
+    public PresignedUrlException(final String message) {
+        super(message);
+    }
+
+    public PresignedUrlException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}
