+++ dms-service/src/main/java/de/muenchen/oss/swim/dms/domain/exception/UnknownUseCaseException.java
@@ -1,0 +1,8 @@
+package de.muenchen.oss.swim.dms.domain.exception;
+
+@SuppressWarnings("PMD.MissingSerialVersionUID")
+public class UnknownUseCaseException extends RuntimeException {
+    public UnknownUseCaseException(final String message) {
+        super(message);
+    }
+}
