+++ dms-service/src/main/java/de/muenchen/oss/swim/dms/domain/model/DmsTarget.java
@@ -1,0 +1,10 @@
+package de.muenchen.oss.swim.dms.domain.model;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record DmsTarget(
+        @NotBlank String coo,
+        @NotBlank String userName,
+        String joboe,
+        String jobposition) {
+}
