--- dms-service/src/main/java/de/muenchen/swim/dms/adapter/in/streaming/DmsEventDTO.java
@@ -1,7 +1,0 @@
-package de.muenchen.swim.dms.adapter.in.streaming;
-
-public record DmsEventDTO(
-        String useCase,
-        String presignedUrl,
-        String metadataPresignedUrl) {
-}
