--- dms-service/src/main/java/de/muenchen/swim/dms/domain/exception/DmsException.java
@@ -1,8 +1,0 @@
-package de.muenchen.swim.dms.domain.exception;
-
-@SuppressWarnings("PMD.MissingSerialVersionUID")
-public class DmsException extends RuntimeException {
-    public DmsException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-}
