--- dms-service/src/main/java/de/muenchen/swim/dms/domain/exception/MetadataException.java
@@ -1,12 +1,0 @@
-package de.muenchen.swim.dms.domain.exception;
-
-@SuppressWarnings("PMD.MissingSerialVersionUID")
-public class MetadataException extends RuntimeException {
-    public MetadataException(final String message) {
-        super(message);
-    }
-
-    public MetadataException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-}
