--- dms-service/src/main/java/de/muenchen/swim/dms/domain/exception/PresignedUrlException.java
@@ -1,12 +1,0 @@
-package de.muenchen.swim.dms.domain.exception;
-
-@SuppressWarnings("PMD.MissingSerialVersionUID")
-public class PresignedUrlException extends RuntimeException {
-    public PresignedUrlException(final String message) {
-        super(message);
-    }
-
-    public PresignedUrlException(final String message, final Throwable cause) {
-        super(message, cause);
-    }
-}
