--- dms-service/src/main/java/de/muenchen/swim/dms/domain/exception/UnknownUseCaseException.java
@@ -1,8 +1,0 @@
-package de.muenchen.swim.dms.domain.exception;
-
-@SuppressWarnings("PMD.MissingSerialVersionUID")
-public class UnknownUseCaseException extends RuntimeException {
-    public UnknownUseCaseException(final String message) {
-        super(message);
-    }
-}
