--- dms-service/src/test/java/de/muenchen/swim/dms/TestConstants.java
@@ -1,16 +1,0 @@
-package de.muenchen.swim.dms;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@SuppressWarnings({ "PMD.TestClassWithoutTestCases", "PMD.DataClass" })
-public final class TestConstants {
-
-    public static final String SPRING_TEST_PROFILE = "test";
-
-    public static final String SPRING_NO_SECURITY_PROFILE = "no-security";
-
-    public static final String SPRING_JSON_LOGGING_PROFILE = "json-logging";
-
-}
